pyproject.toml
[project]
name = "yt-bar"
version = "0.1.0"
description = "macOS menu bar music streamer using yt-dlp"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.4",
    "pyobjc-framework-AVFoundation>=12.0",
    "py2app>=0.28.10",
    "rumps>=0.4.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = ["I"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
